found müller's paper!
I was first introduced to the article through the CDVANDT collection [1]. I was amazed by the quality of the images on the webpage. I tried to hunt it down but couldn’t find it. I half-assumed the figures were created by the late A.O.D. himself. A few days ago, I came across the same paper being referenced in a wonderful video on the V-2 by Blue Paw Print on YouTube [2]. I searched again and finally discovered it was a chapter from [3], which was readily available online.

References:
 1. https://www.cdvandt.org/v2-rocket-control.htm
 2. Inside the V2 Nazi Wonder Weapon: https://www.youtube.com/watch?v=jePOQuFTXy4&t
 3. AGARD - History of German Guided Missiles Development (1956)