complete the text of Müller's paper
Wow! This is a great paper. The level of detail in the design and implementation of the control system is exactly what I was looking for. It's 4:29 AM, and I think I'll go to sleep for now. Maybe tomorrow, I’ll finish working through the equations. I've emailed the cdvandt team to ask if they can send me a high-quality scan of the images from the paper, because I know AOD had a copy of the book. Some of the images he had uploaded to the website were amazing compared to the digital version I currently have.