add figures on müller's paper
I have added the images cropped from a copy of the article that I found online. They are not as high quality as the images on cdvandt.org. However, the site does not contain all the images from the paper. It suggests that AOD owned a copy of the book. I have emailed the cdvandt team to ask if they can share high-quality scans of the images with me. If I receive them, I will replace the ones I am currently using.

My next goal is to re-read the paper, identify any errors, and make the necessary corrections.