add KiCad schematic of differentiator
I am yet to fully grasp the workings of a differentiator. I will make my best effort and consult Ullman should any questions arise. The liberal use of transformers, especially the unusual one (unusual to the dumb me who has no experience with analog technology and vacuum tubes) to the left of the LPF, is quite puzzling to me at the moment.